Fix: now correctly parsing s///w flag (#4)
The old code was failing if the filename for the 'w' flag in s///
started with a letter that was also a s/// flag. Example:

    $ sedsed -e 's/foo/bar/w goo.txt'
    s/foo/bar/wg oo.txt

Now the code handling s/// flags first deals with all the other flags
and then deal with 'w', which is the only flag that accepts an
argument. Now it works as expected:

    $ sedsed -e 's/foo/bar/w goo.txt'
    s/foo/bar/w goo.txt

The test templates were updated, since they incorrectly displayed the
old buggy behavior.

Fixes #4