test: Always fail when extra files are found
The tests_git_status function detects when an unexpected file appears
in the test folder and set the script's error code to 1 in that case.

It also calls tests_clean_up, so with this change the old behavior
is kept and we add a new check.

The other tests already use that function. Those two were the only
outliers.