test: Remove comment after 'w' command
This is a syntax gotcha. From GNU sed documentation:

The r,R,w,W commands parse the filename until end of the line. If
whitespace, comments or semicolons are found, they will be included
in the filename, leading to unexpected results.

So adding that comment after the s///w flag was a mistake. It will
be considered part of the filename. This commit fixes the relevant
test cases.

From the html and token tests fixed in this commit, we can see
that sedsed actually consider that ending a real comment. Issue
https://github.com/aureliojargas/sedsed/issues/29 was opened to
fix sedsed code to correctly parse that.