Rework silly far-right positioned comments
For some reason back in 2001 I've found that it was a good idea.

Now prefix those comments with only two spaces, or convert them to
full line comments.

Note that this was not a total cleanup. There are still some old
aligned comments that were left untouched because I judged they
improve the readability being like that.