WIP - Add support for partial comments
It seems like it is too much trouble, too intrusive in the original
parser.

Even with all this code, I cannot preserve whitespaces before the
'#' char in partial comments, so sedsed would have to add a fixed
amount (as it does today, it's ugly).

At least note that in the test suite results, it fixed a bug in the
original sedsed parser where a partial comment after { gets
duplicated as also a normal comment.

Maybe let it for the future.