gnused.py: Save s///w filename as cmd_subst.outf.name
Previously the filename was saved together with the 'w' flag in the
flags list. Now the flags list only contains the flags themselves
and the s///w filename is saved into the cmd_subst.outf.name property.

It seems less hacky this way, and now we are in line with gsed, which
also uses 'outf' for the s///w file data.