Add setup.py and other package-related changes (#58)
New `setup.py` and `setup.cfg` to generate the package to be used in
pypi.

In `sedsed.py` a new `entrypoint()` function was created to be the
entrypoint for the `sedsed` executable generated by `setup.py`.

New `requirements-pkg.txt` file to list the dependencies necessary to
build the sedsed package.

Removed `requirements.txt`, since the dependency is now set in the
`install_requires` attribute in `setup.py`.

New topic "Packaging" in CONTRIBUTING.md file.

Updated install instructions in README.md, now installing the package
is the official procedure.

Travis CI config updated to install the sedsed package locally.